Fixing: PFWORKBIN-302 and PFWORKBIN-445
Region (type) and Region (variable) have been adjusted so they don't have the same name - This is a requirement for android builds.
A bunch of calls to memcpy() and strncpy() have been replaced with sprintf(), which allows the Android build to compile.

These fixes are for Cocos, but they're all shared code between WindowsSDK and CocosSDK, so both get submitted.


Former-commit-id: 85c676254a2ea3d8dab72613ed59d6099eae1b5f